WIP: Undetected Elements
The reason why there is unexpected errors sometimes when runnning quickie which shows an element is undetected or cannot call because it is null expression and as such -> is because the program runs while executing commands while the chrome driver is slow booting or still loading due to slow server or internet connection.

meaning, the program is shooting into the air because the target is still on delivery.

more to be followed...