WIP: hiding windows
quickiePow: hidden
quickiePass: hidden

flow: quickiePow is already hidden within the StartProcess as an argument. quickiePass is only hidden if quickiePow is done executed. quickiePow runs first in hidden, if done. it will execute exit. and then exits the scripts. technicallly does not hide it but essentially does hide it by exiting the script.

more to be followed...