Sucessfully Integrated Notion Google Log-in
WIP: Pow script and py script has been Integrated. but Notion log in is very flaky, unpredictable with errors. i'm making a retry button as that pops up on the log in process and will appear until log in is sucessful, otherwise retry button will re boot the log in.

more to be followed...