WIP: hashing , crypting and decrypting
Added: Hashing Password on login form finished.
WIP: register Form for emails, hashing is a one way. i need the emails credentials to be extracted from the csv file. entered as a hashed password would result an error.

crypting and decrypting method is work in progress within the user register for emails and email password

more to be followed...